+++ .travis.yml
@@ -1,9 +1,6 @@
 language: node_js
 services: mongodb
 node_js:
-  - "4"
-  - "5"
-  - "6"
   - "7"
 after_success:
   - 'cat coverage/lcov.info | ./node_modules/.bin/coveralls'
