+++ package.json
@@ -4,11 +4,17 @@
     "description": "",
     "main": "index.js",
     "scripts": {
-        "start": "cross-env NODE_ENV=production babel-node index.js | bunyan",
-        "start:development": "cross-env NODE_ENV=development nodemon index.js --exec babel-node --ignore app/views | bunyan",
-        "start:background:imports": "cross-env NODE_ENV=production babel-node background/imports.js | bunyan",
+        "clean": "rm -rf build && mkdir build",
+        "watch": "babel --watch src --out-dir lib --source-maps inline --copy-files",
+        "build": "babel src --out-dir lib --source-maps inline --copy-files",
+        "imports": "cross-env NODE_ENV=production ./cli.js --background-imports",
         "test": "xo && nyc ava"
     },
+    "bin": "cli.js",
+    "files": [
+        "lib/index.js",
+        "cli.js"
+    ],
     "engines": {
         "node": ">=4.0.0"
     },
@@ -27,6 +33,7 @@
         "app-module-path": "^1.0.4",
         "async": "^1.5.0",
         "babel-cli": "^6.24.1",
+        "babel-plugin-transform-runtime": "^6.23.0",
         "babel-polyfill": "^6.23.0",
         "babel-preset-es2015": "^6.24.1",
         "babel-preset-stage-0": "^6.24.1",
@@ -46,6 +53,7 @@
         "http-errors": "^1.6.1",
         "line-by-line": "^0.1.4",
         "loud-rejection": "^1.6.0",
+        "meow": "^3.7.0",
         "method-override": "^2.3.3",
         "mongodb": "^2.1.18",
         "mongoose": "^4.5.0",
@@ -54,6 +62,7 @@
         "parse-torrent": "^5.7.3",
         "pug": "^2.0.0-rc.1",
         "underscore": "^1.8.3",
+        "update-notifier": "^2.2.0",
         "xml2js-es6-promise": "^1.1.1"
     },
     "devDependencies": {
@@ -86,24 +95,17 @@
         "envs": [
             "node",
             "es6"
-        ]
-    },
-    "babel": {
-        "presets": [
-            "es2015",
-            "stage-0",
-            "@ava/transform-test-files"
         ],
-        "plugins": [
-            "transform-regenerator"
+        "ignores": [
+            "lib/**"
         ]
     },
     "ava": {
         "babel": "inherit",
         "require": [
-            "babel-register",
-            "babel-polyfill"
-        ]
+             "babel-core/register",
+             "babel-polyfill"
+         ]
     },
     "nyc": {
         "reporter": [
