+++ src/public/js/core.js
