+++ test/utils.js
@@ -1,5 +1,5 @@
 import test from 'ava';
-import {slugify} from '../utils';
+import {slugify} from '../src/utils';
 
 test('slugify " UnKNown ------ TeST " should return "unknown-test"', t => {
     t.is(slugify(' UnKNown ------ TeST '), 'unknown-test');
